P1 and P2 sync bugfix
YAYY P2's UI now live updates the current instruction, lives, and score. I tweaked with a lot of things but the key change that finally got it to work was adding a listener socket.on for players in game.js--it seemed that when backend emitted thru io.emit, we were missing a socket.on statement in game.js, thus not receiving that info and causing P2 UI to not update.

Not sure why P1's was fine but not P2's, but eager to discuss with Amy later. For now, player's UIs are synced, although the debugging statements on P2 are still stale/not updating... but the actual game components (instruction, lives, score) ARE